Project3/requirements.txt
# 该脚本的作用：定义Python依赖包
Flask==2.3.3
Flask-CORS==4.0.0
pymysql==1.1.0
redis==4.6.0
gunicorn==21.2.0
python-dotenv==1.0.0
